Update popup.html
introduced a navigation bar with a list for different tabs including 'Home' and 'Settings'. The status text is now placed to the left of the switch. The font has been changed to IBM Plex Sans, which is one of the fonts Reddit uses. I've also added an img element at the top for the logo.