Update popup.html
changed the body background to white, the title color to Reddit red, and all other text to black. I've also added a CSS class to visually indicate which tab is selected. This is done by changing the text color to Reddit red and making it bold.

HTML: wrapped the switch control in a div and added another div for the settings page. Both divs have a class "page", and only the one with class "active" will be displayed.